+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- Renamed Money#cents to Money#fractional. Money#cents can still be used as a synonym and equivalent of Money#fractional.
 - Added Money.new_with_amount and Money#amount. Money.new_with_dollars and Money#dollars remain as synonyms.
 - Calling Bank::Base.instance doesn't make Bank::VariableExchange.instance
   return Bank::Base.instance anymore (semaperepelitsa)